Add files via upload
The objective of this project is to share the immense advantage of a technology with those persons who are in a great need of this. Globally  atleast 2.2 billion people have a vision impairment or blindness so the aim of this project is to make their life easier by providing them with this smart gloves. This gloves can tell us if their any obstacles by vibrations and gps tracker that tracks the record of the road and will also tell if the person is lost or not.